Add an architecture overview to the developer guide
This is the beginning of a rough overview of CrateDB's architecture. It is only
a starting point and I hope we can expand it over time. The goal is that people
new to the project can get an idea of what technology we use and how the project
is structured.